ApiPython/requirements.txt
Flask
redis
mysql-connector-python